Refactoring of Json in/out operations
