+++ core/src/main/scala/stainless/MainHelpers.scala
@@ -2,12 +2,13 @@
 
 package stainless
 
-import java.io.{ File, PrintWriter }
+import utils.JsonUtils
+
+import java.io.File
 import java.util.concurrent.{ ExecutorService, Executors }
 
 import org.json4s.JsonAST.JObject
 import org.json4s.JsonDSL._
-import org.json4s.native.JsonMethods._
 
 object MainHelpers {
 
@@ -134,9 +135,7 @@ trait MainHelpers extends inox.MainHelpers {
   private def exportJson(reports: Seq[AbstractReport[_]], file: String): Unit = {
     val subs = reports map { r => JObject(r.name -> r.emitJson) }
     val json = if (subs.isEmpty) JObject() else subs reduce { _ ~ _ }
-    val string = pretty(render(json))
-    val pw = new PrintWriter(new File(file))
-    try pw.write(string) finally pw.close()
+    JsonUtils.writeFile(new File(file), json)
   }
 
 }
